requirements.txt
opencv-python-headless
numpy

python-dotenv

openai
